# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exemple sms",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "description": "Signup by name, email, password and confirme your password",
                "operationId": "authSignup",
                "requestBody": {
                    "description": "User signup",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ali Abdoul"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "User login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "User logout",
                "description": "User logout",
                "operationId": "authLogout",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/send-message": {
            "post": {
                "tags": [
                    "message"
                ],
                "summary": "Send message",
                "description": "Message sending",
                "operationId": "4273ee5d0c985704ef2bf3706bcd3433",
                "requestBody": {
                    "description": "Send message",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contact",
                                    "message",
                                    "sender_id"
                                ],
                                "properties": {
                                    "contact": {
                                        "type": "string",
                                        "example": "+90 545 386 83 09"
                                    },
                                    "message": {
                                        "type": "text",
                                        "format": "text"
                                    },
                                    "sender_id": {
                                        "type": "integer",
                                        "format": "user_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/list-message-sent": {
            "get": {
                "tags": [
                    "message"
                ],
                "summary": "Get messages",
                "description": "Get all messages sent",
                "operationId": "316573cd6ebbdc6aa16afd4e25f7c56b",
                "requestBody": {
                    "description": "Get all message",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/list-message-sent-by-date/{date}": {
            "get": {
                "tags": [
                    "message"
                ],
                "summary": "Get messages",
                "description": "Get all messages sent on one date",
                "operationId": "44abe3cdfedeabf712ac16db1320f8ce",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/get-message-details/{id}": {
            "get": {
                "tags": [
                    "message"
                ],
                "summary": "Get one message details",
                "description": "Get all details for one message",
                "operationId": "ee2a556afebdbd910d047e13df7dc1af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}